requirements.txt
asgiref==3.8.1
Django==5.1.6
django-ckeditor==6.7.2
django-js-asset==3.0.1
django-taggit==6.1.0
pillow==11.1.0
psycopg2==2.9.10
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.1
gunicorn==23.0.0
whitenoise==6.6.0  